# src/app/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "QR Generator API",
    "version": "1.0.0",
    "description": "REST API for managing QR codes, scans, and webhooks",
    "contact": {
      "name": "API Support",
      "url": "https://your-domain.com/support"
    }
  },
  "servers": [
    {
      "url": "https://your-domain.com/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "QRCode": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "url": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "foregroundColor": { "type": "string" },
          "backgroundColor": { "type": "string" },
          "dotType": { "type": "string", "enum": ["square", "rounded", "dots", "classy", "classy-rounded", "extra-rounded"] },
          "cornerType": { "type": "string", "enum": ["square", "rounded", "dot", "extra-rounded"] },
          "logoUrl": { "type": "string" },
          "hasWatermark": { "type": "boolean" },
          "isDynamic": { "type": "boolean" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "maxScans": { "type": "integer" },
          "redirectUrl": { "type": "string" },
          "webhookUrl": { "type": "string" },
          "scanCount": { "type": "integer" },
          "downloadCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/qr-codes": {
      "get": {
        "summary": "List QR codes",
        "tags": ["QR Codes"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 }, "description": "Number of results to return" },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 }, "description": "Number of results to skip" },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search query" }
        ],
        "responses": {
          "200": {
            "description": "List of QR codes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "qrCodes": { "type": "array", "items": { "$ref": "#/components/schemas/QRCode" } },
                    "total": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "summary": "Create QR code",
        "tags": ["QR Codes"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": { "type": "string" },
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "foregroundColor": { "type": "string" },
                  "backgroundColor": { "type": "string" },
                  "dotType": { "type": "string" },
                  "cornerType": { "type": "string" },
                  "logoUrl": { "type": "string" },
                  "hasWatermark": { "type": "boolean" },
                  "isDynamic": { "type": "boolean" },
                  "expiresAt": { "type": "string", "format": "date-time" },
                  "maxScans": { "type": "integer" },
                  "redirectUrl": { "type": "string" },
                  "webhookUrl": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "QR code created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "qrCode": { "$ref": "#/components/schemas/QRCode" }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}